+++ src/templates/speciality.js
@@ -6,7 +6,7 @@ import IntroSection from '../components/Speciality/Intro'
 import ProjectsSection from '../components/Speciality/Projects'
 import TrainingSection from '../components/Speciality/Training'
 import CommunitySection from '../components/Speciality/Community'
-import EventSection from '../components/Speciality/Events'
+import EventSection from '../components/Common/Events'
 import TalksSection from '../components/Speciality/Talks'
 import GetInTouch from '../components/Common/GetInTouch'
 import TutorialsSection from '../components/Speciality/Tutorials'
